3gram重みを小数設定可能に変更
- char3gramWeight を C#、kw-uni、NgramAnalyzer で double として扱う

- NgramAnalyzer の起動オプションを FLOAT 型に変更する

+++ AyaoriHIME/Settings.cs
@@ -258,7 +258,7 @@ namespace KanchokuWS
         public static int NgramCostFactor { get; set; } = 5;
 
         /// <summary>文字3-gram言語モデルの重み</summary>
-        public static int Char3gramWeight { get; set; } = 1;
+        public static double Char3gramWeight { get; set; } = 1.0;
 
         /// <summary>文字3-gram窓末尾が漢字の場合のコスト減衰率</summary>
         public static double Char3gramTailKanjiCostDecayRate { get; set; } = 0.5;
@@ -1967,9 +1967,9 @@ namespace KanchokuWS
             MorphNonTerminalCost = addDecoderSetting("morphNonTerminalCost", 10000);            // 非終端形態素の単語コスト
             AnalyzeMorphLen = addDecoderSetting("analyzeMorphLen", 10);                         // 形態素解析を行う際の最小形態素長
             NgramCostFactor = addDecoderSetting("ngramCostFactor", 5);                          // 形態素コストに対するNgramコストの係数
-            Char3gramWeight = addDecoderSetting("char3gramWeight", 1, 0);                       // 文字3-gram言語モデルの重み
+            Char3gramWeight = addDecoderSetting("char3gramWeight", 1.0, 0.0);                   // 文字3-gram言語モデルによるコストの重み係数
             Char3gramTailKanjiCostDecayRate = addDecoderSetting("char3gramTailKanjiCostDecayRate", 0.5, 0.0);
-            Char4gramWeight = addDecoderSetting("char4gramWeight", 0.0, 0.0);                   // 文字4-gramボーナスの係数
+            Char4gramWeight = addDecoderSetting("char4gramWeight", 0.0, 0.0);                   // 文字4-gramボーナスの係数 (正値にすると Char3gramひらがなのコストは 0 扱いになる)
             NgramMaxBonusPoint = addDecoderSetting("ngramMaxBonusPoint", 25);                   // Ngramに与えるボーナスポイントの最大値
             NgramBonusPointFactor = addDecoderSetting("ngramBonusPointFactor", 250);            // 嵩上げされたNgramに与えるボーナスの係数
             NgramManualSelectDelta = addDecoderSetting("ngramManualSelectDelta", 10);           // 候補選択によるNgramカウントの変動幅
